Trailing zeroes in factorial
For an integer n find number of trailing zeroes in n! . 

Input:
The first line contains an integer 'T' denoting the total number of test cases. In each test cases, it contains an integer 'N'.


Output:
In each seperate line output the answer to the problem.


Constraints:
1<=T<=100
1<=N<=1000


Example:
Input:
1
9

Output:
1